+++ examples/simple_rule.rs
@@ -34,5 +34,7 @@ async fn main() {
 
     // Evaluate the ruleset on the input data and check if the rule returns
     // `false`
-    assert_eq!(ruleset.evaluate(&facts).await.unwrap(), vec![false.into()]);
+    for outcome in ruleset.evaluate(&facts).await.unwrap() {
+        assert_eq!(outcome.value.unwrap(), false.into());
+    }
 }