+++ src/lib.rs
@@ -37,7 +37,9 @@
 //! let facts = Data { age: 16 };
 //! // Evaluate the ruleset on the input data and check if the rule returns
 //! // `false`
-//! assert_eq!(ruleset.evaluate(&facts).await.unwrap(), vec![false.into()]);
+//! for outcome in ruleset.evaluate(&facts).await.unwrap() {
+//!     assert_eq!(outcome.value.unwrap(), false.into());
+//! }
 //! # })
 //!  ```
 //!
@@ -90,7 +92,9 @@
 //! let facts = Data { age: 16 };
 //! // Evaluate the ruleset on the input data and check if the rule returns
 //! // `true`
-//! assert_eq!(ruleset.evaluate(&facts).await.unwrap(), vec![true.into()]);
+//! for outcome in ruleset.evaluate(&facts).await.unwrap() {
+//!     assert_eq!(outcome.value.unwrap(), true.into());
+//! }
 //! # })
 //!  ```